Factor tinyMCE specific rendering into mixin.
